backend/requirements.txt
langchain
langgraph
pydantic
python-dotenv
httpx
beautifulsoup4
google-generativeai
google-genai
tavily-python
google-search-results
questionary
wikipedia
yfinance
huggingface_hub
numpy
fastapi
uvicorn
motor
passlib[bcrypt]
bcrypt>=4.0.0
python-jose[cryptography]
python-multipart
email-validator
Pillow
asyncpraw
apscheduler
telethon
geopy